Refactor: async agent construction, MCP integration
Major refactor to use async agent factories throughout. Updated agent orchestration, workflows, and tool registry for async signatures. Improved MCP tool integration for runtime safety and non-blocking access. Updated tests, removed obsolete code, and clarified solution structure and dependency rules. Documentation and project files revised to match new architecture. Minor fixes for formatting, namespaces, and comments.

+++ projects/SentinelCore.Tests/CaseFlow/EvidenceStoreTests.cs
@@ -2,13 +2,15 @@
 // Project:   SentinelCore.Tests
 // File:         EvidenceStoreTests.cs
 // Author: Kyle L. Crowder
-// Build Num:  080801
+// Build Num:  081312
 
 
 
-using SentinelCore.CaseFlow;
+using Microsoft.EntityFrameworkCore;
+
+using SentinelCore.Abstractions;
+using SentinelCore.CaseEngine;
 using SentinelCore.CaseFlowEngine.Persistence;
-using SentinelCore.Infrastructure.Persistence;
 
 
 
@@ -43,7 +45,7 @@ public sealed class EvidenceStoreTests
         CaseEntity caseEntity = new()
         {
                 CaseId = caseGuid,
-                Status = CaseStatus.Open,
+                Status = (int)CaseStatus.Open,
                 CreatedAt = DateTime.Now,
                 UpdatedAt = DateTime.Now,
                 InitiatingSignal = 0
@@ -157,7 +159,7 @@ public sealed class EvidenceStoreTests
         CaseEntity caseEntity = new()
         {
                 CaseId = caseGuid,
-                Status = CaseStatus.Open,
+                Status = (int)CaseStatus.Open,
                 CreatedAt = DateTime.Now,
                 UpdatedAt = DateTime.Now,
                 InitiatingSignal = 0
